+++ spec/grover/processor_spec.rb
@@ -408,7 +408,7 @@ describe Grover::Processor do
 
               <script>
                 setTimeout(function() {
-                  document.body.innerHTML = '<h1>Hey there</h1>'
+                  document.body.innerHTML = '<h1>Hey there</h1>';
                 }, 100);
               </script>
             </html>
@@ -418,6 +418,33 @@ describe Grover::Processor do
         let(:date) { Date.today.strftime '%-m/%-d/%Y' }
 
         it { expect(pdf_text_content).to eq "#{date} Hey there http://www.example.net/foo/bar 1/1" }
+      end
+
+      context 'when wait for selector option is specified with options' do
+        let(:url_or_html) do
+          <<-HTML
+            <html>
+              <body>
+                <p id="loading">Loading</p>
+              </body>
+
+              <script>
+                setTimeout(function() {
+                  document.getElementById('loading').remove()
+                }, 100);
+              </script>
+            </html>
+          HTML
+        end
+        let(:options) do
+          basic_header_footer_options.merge(
+            'waitForSelector' => '#loading',
+            'waitForSelectorOptions' => { 'hidden' => true }
+          )
+        end
+        let(:date) { Date.today.strftime '%-m/%-d/%Y' }
+
+        it { expect(pdf_text_content).to eq "#{date} http://www.example.net/foo/bar 1/1" }
       end
     end