+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - [#63](https://github.com/Studiosity/grover/pull/63) Ensure cookies from incoming request are passed to Grover via Middleware ([@braindeaf][])
+- [#64](https://github.com/Studiosity/grover/pull/64) Add waitForSelector support ([@andmcgregor][])
 
 ## [0.12.1](releases/tag/v0.12.1) - 2020-05-12
 ### Fixed